Auth and Guest middleware

+++ routes.php
@@ -2,18 +2,18 @@
 
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
 $router->get('/listing/{id}', 'ListingController@show');
-$router->get('/listing/edit/{id}', 'ListingController@edit');
+$router->get('/listing/edit/{id}', 'ListingController@edit', ['auth']);
 
-$router->post('/listings', 'ListingController@store');
+$router->post('/listings', 'ListingController@store', ['auth']);
 
-$router->put('/listing/{id}', 'ListingController@update');
-$router->delete('/listing/{id}', 'ListingController@destroy');
+$router->put('/listing/{id}', 'ListingController@update', ['auth']);
+$router->delete('/listing/{id}', 'ListingController@destroy', ['auth']);
 
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
 
-$router->post('/auth/register', 'UserController@store');
-$router->post('/auth/login', 'UserController@authorize');
-$router->post('/auth/logout', 'UserController@destroy');
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/login', 'UserController@authorize', ['guest']);
+$router->post('/auth/logout', 'UserController@destroy', ['auth']);
